Change tokenToText APIs into tokensToText. With a plural.
The assumption that converting each token individually text and then
concatenate does not seem to hold.

I didn't see an API for llama.cpp that does this; it's got the function
to turn into pieces but no converter to do it in bulk.

Some fixes to the Python code as well, respects local files and trust
remote now.